Door Opening Mechanics for White Room
Door Opening Mechanics for the White Room are set up but I am running
into an issue with the black wall/door in the Red Room that hasn't be
changed into a door yet. As it stands right now, it is impossible to
even enter the red room because of said door/wall. Whereas the door
opening for the other doors is just a modification of the code that is
needed.